server/worker/sentiment_analysis/requirements.txt
pika==1.3.2
configparser
transformers==4.18.0
torch>=2.0.0
numpy>=1.22.3
python-dotenv==1.1.0